pyproject.toml
[project]
name = "obs-utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "textual>=6.5.0",
    "pyserial-asyncio>=0.6",
    "pyudev>=0.24.4",
    "pyyaml>=6.0.3",
    "pyzmq>=27.1.0",
    "rich>=14.2.0",
    "uvloop>=0.22.1",
    "astrix[jax]",
    "numpy>=2.3.4",
    "opencv-contrib-python>=4.11.0.86",
    "pyqt6>=6.10.0",
    "pyqtgraph>=0.13.7",
    "vmbpy[numpy,opencv]",
    "pandas>=2.3.3",
    "jax>=0.8.0",
    "astropy>=7.1.1",
    "prompt-toolkit>=3.0.52",
]

[tool.setuptools]
packages = ["obs_utils", "obs_tui", "obs_certus", "obs_encoders", "obs_display", "obs_cameras", "obs_controller", "obs_target"]

[project.scripts]
obs-tui = "obs_tui.__main__:main"
obs-show-devices = "obs_utils.discovery:print_search"
obs-certus = "obs_certus.__main__:main"
obs-encoders = "obs_encoders.__main__:main"

[tool.uv.sources]
vmbpy = { url = "https://github.com/alliedvision/VmbPy/releases/download/1.2.0/vmbpy-1.2.0-py3-none-manylinux_2_27_x86_64.whl" }
astrix = { git = "https://github.com/andrewjlock/astrix.git" }

[tool.basedpyright]
typeCheckingMode = "standard"
reportIgnoreCommentWithoutRule = false

[tool.setuptools.package-data]
obs_certus = ["cpp/certus/stream_packets"]
